requirements.txt
beautifulsoup4==4.12.3
gradio==5.9.1
langchain==0.3.13
langchain_anthropic==0.3.1
langchain_community==0.3.13
langchain_core==0.3.28
langchain_experimental==0.3.4
langchain_mongodb==0.3.0
langchain_ollama==0.2.2
langchain_openai==0.2.14
langchain_together==0.2.0
Pillow==11.0.0
pillow_heif==0.20.0
pydantic==2.10.4
pymupdf4llm==0.0.17
python-dotenv==1.0.1
Requests==2.32.3
wikipedia==1.4.0
xmltodict==0.14.2
numexpr==2.10.2
arxiv==2.1.3
pyowm==3.3.0
setuptools==75.6.0

